requirements.txt
# Основные библиотеки
Flask==2.3.3
python-dotenv==1.0.0

# Библиотеки для работы с Telegram API
Telethon==1.30.3

# Библиотеки для обработки данных
pandas==2.1.0
numpy==1.25.2
nltk==3.8.1

# Библиотеки для работы с OpenAI API
openai==0.28.0

# Библиотеки для форматирования и экспорта отчетов
markdown==3.4.4
pdfkit==1.0.0

# Дополнительные зависимости
requests==2.31.0
python-dateutil==2.8.2